fix: Add on/off/toggle methods for advanced sharing
This fixes and issue with Mobile Control not triggering a change of mode if the mode is  not changed using the Mobile Control App.

Also added a path, `/room/room1/routing/advancedSharingMode/toggle` that will pulse the toggle join when received. This path is NOT currently supported in the Mobile Control app.