feat: RKST-95 send list of MC touch panels on websocket join
In order to allow touch panels to communicate with the processor through the MC API, the touch panels need a permanent join token, similar to how the direct server configuraiton works.
Sending the list of touch panels to the API allows the API to create those tokens and send them back to the plugin to send to the panel with the correct URL for the mobile control application.