fix: call press hold action in timer methods
Previously, calling the action in both the `HandlePressAndHold` method and `StopTimer` was resulting in dual calls when a button was released. The calls to the action in `HandlePressAndHold` have been removed in favor of calling the action with `true` in `AddTimer` and calling the action with `false` in `StopTimer`. This results in the expected calls to the methods happening.